Refactor: Extract backspace logic to dedicated handler (#480)
